+++ slave/conn_tracker.go
@@ -1,4 +1,4 @@
-package server
+package slave
 
 import (
 	"net"
